feat: Basic library functionality
Ported over much of my work from the falling sand simulator, and then
almost rewrote everything. I've kept the `Grid` type which is the main
entry point, but now it contains `Cell` structs instead of a `bool`,
which allows for color customisation and what two characters to display
in a cell. The draw method leaves much to be desired, it could probably
implement double buffering for more intense rendering to not slow down
everything. Either this or only updating cells that have changed,
essentially atmoic editing which I implemented for the sand falling
simulation. I'd like to implement a get cell click function for the
mouse into this library too, so that people who use this library don't
have to account for the weird terminal cell width to pixel width (/2)