fix: implemented a changes buffer to improve drawing performance
Changes to the grid are now stored in a vec containing `CellChange`
structs. If this changes vec is not empty, then the `draw` calls a
function that only draws the changes onto the grid. If the changes vec
is empty, the whole grid is drawn. This seems like it dramatically
increases performance from testing, as the canvas example went from
lagging behind heaps and maxing out the CPU, to barely using anything